+++ app/model/knock/auth_token.rb
@@ -17,8 +17,21 @@ module Knock
       end
     end
 
-    def current_user
-      @current_user ||= Knock.current_user_from_token.call @payload
+    def entity_for entity_class
+      if entity_class.respond_to? :from_token_payload
+        entity_class.from_token_payload @payload
+      else
+        if entity_class.to_s == "User" && Knock.respond_to?(:current_user_from_token)
+          warn "[DEPRECATION]: `Knock.current_user_from_token` is deprecated. Please implement `User.from_token_payload` instead."
+          Knock.current_user_from_token.call @payload
+        else
+          entity_class.find @payload['sub']
+        end
+      end
+    end
+
+    def to_json options = {}
+      {jwt: @token}.to_json
     end
 
   private