+++ lib/generators/templates/knock.rb
@@ -1,5 +1,8 @@
 Knock.setup do |config|
 
+  ## [DEPRECATED]
+  ## This is deprecated in favor of `User.from_token_request`.
+  ##
   ## User handle attribute
   ## ---------------------
   ##
@@ -8,6 +11,9 @@ Knock.setup do |config|
   ## Default:
   # config.handle_attr = :email
 
+  ## [DEPRECATED]
+  ## This is deprecated in favor of `User.from_token_request`.
+  ##
   ## Current user retrieval from handle when signing in
   ## --------------------------------------------------
   ##
@@ -25,6 +31,9 @@ Knock.setup do |config|
   ## Default:
   # config.current_user_from_handle = -> (handle) { User.find_by! Knock.handle_attr => handle }
 
+  ## [DEPRECATED]
+  ## This is depreacted in favor of `User.from_token_payload`.
+  ##
   ## Current user retrieval when validating token
   ## --------------------------------------------
   ##
@@ -92,8 +101,7 @@ Knock.setup do |config|
   ## Exception Class
   ## ---------------
   ##
-  ## Configure the Exception to be used (raised and rescued) for User Not Found.
-  ## note: change this if ActiveRecord is not being used.
+  ## Configure the exception to be used when user cannot be found.
   ##
   ## Default:
   # config.not_found_exception_class_name = 'ActiveRecord::RecordNotFound'