+++ test/dummy/app/controllers/vendor_protected_controller.rb
@@ -1,0 +1,11 @@
+class VendorProtectedController < ApplicationController
+  before_action :authenticate_vendor, only: [:index]
+  before_action :some_missing_method, only: [:show]
+
+  def index
+    head :ok
+  end
+
+  def show
+  end
+end