+++ test/dummy/app/controllers/vendor_token_controller.rb
@@ -1,0 +1,2 @@
+class VendorTokenController < Knock::AuthTokenController
+end