+++ test/dummy/app/models/vendor.rb
@@ -1,0 +1,3 @@
+class Vendor < ActiveRecord::Base
+  has_secure_password
+end