+++ test/dummy/db/migrate/20160519075733_create_admins.rb
@@ -1,0 +1,10 @@
+class CreateAdmins < ActiveRecord::Migration
+  def change
+    create_table :admins do |t|
+      t.string :email
+      t.string :password_digest
+
+      t.timestamps null: false
+    end
+  end
+end