+++ test/model/knock/auth_token_test.rb
@@ -4,24 +4,25 @@ require 'timecop'
 
 module Knock
   class AuthTokenTest < ActiveSupport::TestCase
+    setup do
+      key = Knock.token_secret_signature_key.call
+      @token = JWT.encode({sub: '1'}, key, 'HS256')
+    end
+
     test "verify algorithm" do
       Knock.token_signature_algorithm = 'RS256'
-      key = Knock.token_secret_signature_key.call
-
-      token = JWT.encode({sub: '1'}, key, 'HS256')
 
       assert_raises(JWT::IncorrectAlgorithm) {
-        AuthToken.new(token: token)
+        AuthToken.new(token: @token)
       }
     end
 
     test "decode RSA encoded tokens" do
-      user = users(:one)
       rsa_private = OpenSSL::PKey::RSA.generate 2048
       Knock.token_public_key = rsa_private.public_key
       Knock.token_signature_algorithm = 'RS256'
 
-      token = JWT.encode({sub: user.id}, rsa_private, 'RS256')
+      token = JWT.encode({sub: "1"}, rsa_private, 'RS256')
 
       assert_nothing_raised { AuthToken.new(token: token) }
     end
@@ -42,10 +43,8 @@ module Knock
       Knock.token_audience = -> { 'bar' }
       key = Knock.token_secret_signature_key.call
 
-      token = JWT.encode({sub: 'foo'}, key, 'HS256')
-
       assert_raises(JWT::InvalidAudError) {
-        AuthToken.new token: token
+        AuthToken.new token: @token
       }
     end
 
@@ -65,6 +64,12 @@ module Knock
       Timecop.travel(10.years.from_now) do
         assert_not Knock::AuthToken.new(token: token).payload.has_key?('exp')
       end
+    end
+
+    test "is serializable" do
+      auth_token = AuthToken.new token: @token
+
+      assert_equal("{\"jwt\":\"#{@token}\"}", auth_token.to_json)
     end
   end
 end