+++ test/test_helper.rb
@@ -1,4 +1,7 @@
+require 'simplecov'
 require "codeclimate-test-reporter"
+
+SimpleCov.start
 CodeClimate::TestReporter.start
 
 # Configure Rails Environment